// src/models/SupplierManagementObject.model.ts
export interface SupplierManagementObject {
    name?: string;
    surname?: string;
    position?: string;
    idd?: string;
    phoneNumber?: string;
    emailAddress?: string;
    name2?: string;
    surname2?: string;
    position2?: string;
    idd2?: string;
    phoneNumber2?: string;
    emailAddress2?: string;
}